feature: metrics 0.14 & ability to flush to disk less

+++ src/schema.rs
@@ -3,7 +3,7 @@ table! {
         id -> BigInt,
         timestamp -> Double,
         key -> Text,
-        value -> BigInt,
+        value -> Double,
     }
 }
